ADD VISUAL STUDIO CODE EDITING SHORTCUTS
In VSCode, sometimes we are in the middle of an autocompleted
line and need to go to the end of the line by pressing "End"
and then the line terminator.
Instead this shortcut traps a double press of the ; key and
automatically inserts a semi-colon at the end of the line.

Same thing for braces, colons ...
This was from a script that I wrote for WebStorm a few years
ago.